requirements.txt
Flask
python-dotenv
PyPDF2
langchain

